# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CaptureSnapshotOut": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaptureItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "resource_id": {
            "title": "Resource Id",
            "type": "integer"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope",
          "source_type",
          "resource_id",
          "database_name",
          "status",
          "attempt",
          "error_message",
          "started_at",
          "finished_at"
        ],
        "title": "CaptureSnapshotOut",
        "type": "object"
      },
      "ContractDataFetchRequest": {
        "properties": {
          "contract_types": {
            "items": {
              "enum": [
                "futures",
                "options"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Contract Types",
            "type": "array"
          },
          "database_name": {
            "title": "Database Name",
            "type": "string"
          },
          "database_resource_id": {
            "title": "Database Resource Id",
            "type": "integer"
          }
        },
        "required": [
          "database_resource_id",
          "database_name",
          "contract_types"
        ],
        "title": "ContractDataFetchRequest",
        "type": "object"
      },
      "ContractDataFileOut": {
        "properties": {
          "checksum": {
            "title": "Checksum",
            "type": "string"
          },
          "contract_type": {
            "title": "Contract Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "database_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Name"
          },
          "database_resource_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Resource Id"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "order_resource_id": {
            "title": "Order Resource Id",
            "type": "integer"
          },
          "preview_rows": {
            "items": {
              "type": "object"
            },
            "title": "Preview Rows",
            "type": "array"
          },
          "quote_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Date"
          },
          "remote_path": {
            "title": "Remote Path",
            "type": "string"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "source_table": {
            "title": "Source Table",
            "type": "string"
          },
          "workflow_node_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Node Id"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "workflow_node_id",
          "order_resource_id",
          "database_resource_id",
          "database_name",
          "contract_type",
          "source_table",
          "filename",
          "remote_path",
          "quote_date",
          "row_count",
          "size",
          "checksum",
          "preview_rows",
          "created_at"
        ],
        "title": "ContractDataFileOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScenarioOut": {
        "properties": {
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_resource_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Default Resource Ids",
            "type": "array"
          },
          "draft_workflow_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Workflow Version Id"
          },
          "expected_artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Expected Artifacts",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "integer"
          },
          "published_workflow_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Workflow Version Id"
          },
          "required_resource_types": {
            "items": {
              "type": "string"
            },
            "title": "Required Resource Types",
            "type": "array"
          },
          "scenario_type": {
            "title": "Scenario Type",
            "type": "string"
          },
          "workflow_status": {
            "title": "Workflow Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plan_id",
          "name",
          "scenario_type",
          "config_version",
          "expected_artifacts",
          "default_resource_ids",
          "required_resource_types",
          "is_enabled",
          "workflow_status",
          "draft_workflow_version_id",
          "published_workflow_version_id",
          "is_archived",
          "created_at"
        ],
        "title": "ScenarioOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowDocumentOut": {
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/WorkflowVersionOut"
          },
          "published_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Version Id"
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioOut"
          },
          "validation_errors": {
            "items": {
              "type": "object"
            },
            "title": "Validation Errors",
            "type": "array"
          }
        },
        "required": [
          "scenario",
          "draft",
          "published_version_id"
        ],
        "title": "WorkflowDocumentOut",
        "type": "object"
      },
      "WorkflowDocumentWrite": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNodeWrite"
            },
            "title": "Nodes",
            "type": "array"
          },
          "resource_ids": {
            "items": {
              "type": "integer"
            },
            "minItems": 1,
            "title": "Resource Ids",
            "type": "array"
          }
        },
        "required": [
          "expected_revision",
          "resource_ids"
        ],
        "title": "WorkflowDocumentWrite",
        "type": "object"
      },
      "WorkflowNodeOut": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "title": "Config",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "node_key": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Node Key",
            "type": "string"
          },
          "node_type": {
            "enum": [
              "server_config",
              "database_config",
              "wiring_confirmation",
              "order_preparation",
              "slnic_start_capture",
              "slnic_stop_capture",
              "slnic_merge_capture",
              "parser_parse"
            ],
            "title": "Node Type",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "node_key",
          "node_type",
          "name",
          "id",
          "position"
        ],
        "title": "WorkflowNodeOut",
        "type": "object"
      },
      "WorkflowNodeWrite": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "title": "Config",
            "type": "object"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "node_key": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Node Key",
            "type": "string"
          },
          "node_type": {
            "enum": [
              "server_config",
              "database_config",
              "wiring_confirmation",
              "order_preparation",
              "slnic_start_capture",
              "slnic_stop_capture",
              "slnic_merge_capture",
              "parser_parse"
            ],
            "title": "Node Type",
            "type": "string"
          }
        },
        "required": [
          "node_key",
          "node_type",
          "name"
        ],
        "title": "WorkflowNodeWrite",
        "type": "object"
      },
      "WorkflowVersionOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNodeOut"
            },
            "title": "Nodes",
            "type": "array"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "published_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published By"
          },
          "resource_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Resource Ids",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version_no": {
            "title": "Version No",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "version_no",
          "status",
          "revision",
          "resource_ids",
          "published_by",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "WorkflowVersionOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "OpenSLT",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scenarios/{scenario_id}/workflow": {
      "get": {
        "operationId": "get_scenario_workflow_api_v1_scenarios__scenario_id__workflow_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Scenario Workflow"
      },
      "put": {
        "operationId": "save_scenario_workflow_api_v1_scenarios__scenario_id__workflow_put",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDocumentWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Scenario Workflow"
      }
    },
    "/api/v1/scenarios/{scenario_id}/workflow/draft": {
      "post": {
        "operationId": "ensure_workflow_draft_api_v1_scenarios__scenario_id__workflow_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ensure Workflow Draft"
      }
    },
    "/api/v1/scenarios/{scenario_id}/workflow/nodes/{node_key}/contract-files": {
      "get": {
        "operationId": "list_contract_files_api_v1_scenarios__scenario_id__workflow_nodes__node_key__contract_files_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "node_key",
            "required": true,
            "schema": {
              "title": "Node Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContractDataFileOut"
                  },
                  "title": "Response List Contract Files Api V1 Scenarios  Scenario Id  Workflow Nodes  Node Key  Contract Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Contract Files"
      }
    },
    "/api/v1/scenarios/{scenario_id}/workflow/nodes/{node_key}/contract-files/fetch": {
      "post": {
        "operationId": "create_contract_files_api_v1_scenarios__scenario_id__workflow_nodes__node_key__contract_files_fetch_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "node_key",
            "required": true,
            "schema": {
              "title": "Node Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractDataFetchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContractDataFileOut"
                  },
                  "title": "Response Create Contract Files Api V1 Scenarios  Scenario Id  Workflow Nodes  Node Key  Contract Files Fetch Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Contract Files"
      }
    },
    "/api/v1/scenarios/{scenario_id}/workflow/nodes/{node_key}/preview": {
      "post": {
        "operationId": "preview_workflow_node_api_v1_scenarios__scenario_id__workflow_nodes__node_key__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "node_key",
            "required": true,
            "schema": {
              "title": "Node Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaptureSnapshotOut"
                  },
                  "title": "Response Preview Workflow Node Api V1 Scenarios  Scenario Id  Workflow Nodes  Node Key  Preview Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Workflow Node"
      }
    },
    "/api/v1/scenarios/{scenario_id}/workflow/publish": {
      "post": {
        "operationId": "publish_scenario_workflow_api_v1_scenarios__scenario_id__workflow_publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Scenario Workflow"
      }
    },
    "/api/v1/scenarios/{scenario_id}/workflow/versions": {
      "get": {
        "operationId": "list_workflow_versions_api_v1_scenarios__scenario_id__workflow_versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowVersionOut"
                  },
                  "title": "Response List Workflow Versions Api V1 Scenarios  Scenario Id  Workflow Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Workflow Versions"
      }
    },
...
  }
}